requirements.txt
disnake
pymongo
dnspython
python-dotenv
asyncio
re
os
